some components refactored
