+++ app/AppState/AddressBookEntry.ts
@@ -1,0 +1,9 @@
+export class AddressBookEntry {
+  label: string;
+  address: string;
+
+  constructor(label: string, address: string) {
+    this.label = label;
+    this.address = address;
+  }
+}
