+++ app/AppState/ConnectedCompanionApp.ts
@@ -1,0 +1,4 @@
+export interface ConnectedCompanionApp {
+  name: string;
+  lastSeen: number;
+}
