+++ app/AppState/ErrorModalData.ts
@@ -1,0 +1,12 @@
+export class ErrorModalData {
+  title: string;
+  body: string;
+
+  modalIsOpen: boolean;
+
+  constructor() {
+    this.modalIsOpen = false;
+    this.title = '';
+    this.body = '';
+  }
+}
