+++ app/AppState/SendJsonToType.ts
@@ -1,0 +1,5 @@
+export interface SendJsonToType {
+  address: string;
+  amount: number;
+  memo?: string;
+}
