+++ app/AppState/SendPageState.ts
@@ -1,0 +1,9 @@
+import { ToAddr } from './ToAddr';
+
+export class SendPageState {
+  toaddr: ToAddr;
+
+  constructor(toaddr: ToAddr) {
+    this.toaddr = toaddr;
+  }
+}
