+++ app/AppState/ServerSelectState.ts
@@ -1,0 +1,3 @@
+export interface ServerSelectState {
+  modalIsOpen: boolean;
+}
