+++ app/AppState/SettingsFileEntry.ts
@@ -1,0 +1,11 @@
+export class SettingsFileEntry {
+  server?: string;
+  language?: string;
+  currency?: string;
+
+  constructor(server: string) {
+    this.server = server;
+    this.language = '';
+    this.currency = '';
+  }
+}
