+++ app/AppState/SyncStatus.ts
@@ -1,0 +1,5 @@
+export interface SyncStatus {
+  inProgress: boolean;
+  progress: number;
+  blocks: string;
+}
