+++ app/AppState/ToAddr.ts
@@ -1,0 +1,15 @@
+export class ToAddr {
+  id: number;
+  to: string;
+  amount: string;
+  amountUSD: string;
+  memo: string;
+
+  constructor(id: number) {
+    this.id = id;
+    this.to = '';
+    this.amount = '';
+    this.amountUSD = '';
+    this.memo = '';
+  }
+}
