+++ app/AppState/TxDetailType.ts
@@ -1,0 +1,5 @@
+export interface TxDetailType {
+  address: string;
+  amount: number;
+  memo: string | null;
+}
