+++ app/AppState/WalletSeed.ts
@@ -1,0 +1,4 @@
+export interface WalletSeed {
+  seed: string;
+  birthday: number;
+}
