+++ app/rpc/index.js
@@ -1,0 +1,3 @@
+import RPC from './RPC';
+
+export default RPC;
