+++ app/types/ColorsType.ts
@@ -1,8 +1,3 @@
-export type ThemeType = {
-  dark: boolean;
-  colors: ColorsType;
-};
-
 export type ColorsType = {
   background: string;
   card: string;
