+++ app/types/ThemeType.ts
@@ -1,0 +1,6 @@
+import { ColorsType } from './ColorsType';
+
+export type ThemeType = {
+  dark: boolean;
+  colors: ColorsType;
+};
