+++ app/types/index.js
@@ -1,0 +1,4 @@
+import ThemeType from './ThemeType';
+import ColorsType from './ColorsType';
+
+export { ThemeType, ColorsType };
