+++ app/uris/SERVER_URI.ts
@@ -1,0 +1,1 @@
+export const SERVER_URI: string[] = ['https://mainnet.lightwalletd.com:9067', 'https://zuul.free2z.cash:9067'];
