+++ app/utils/ZecAmountSplitType.ts
@@ -1,0 +1,4 @@
+export type ZecAmountSplitType = {
+  bigPart: string;
+  smallPart: string;
+};
