+++ app/utils/index.js
@@ -1,0 +1,3 @@
+import Utils from './Utils';
+
+export default Utils;
