+++ components/Info/Info.tsx
@@ -106,7 +106,14 @@ const Info: React.FunctionComponent<InfoProps> = ({ info, closeModal, totalBalan
         </View>
       </ScrollView>
 
-      <View style={{ flexGrow: 1, flexDirection: 'row', justifyContent: 'center', alignItems: 'center', marginVertical: 5 }}>
+      <View
+        style={{
+          flexGrow: 1,
+          flexDirection: 'row',
+          justifyContent: 'center',
+          alignItems: 'center',
+          marginVertical: 5,
+        }}>
         <Button type="Secondary" title={translate('close')} onPress={closeModal} />
       </View>
     </SafeAreaView>
